ref: remove replicated code
