+++ src/client/components/addWallet/ImportWallet.tsx
@@ -110,7 +110,7 @@ const ImportWallet = ({ mnemonic, onSuccess }: ImportWalletProps) => {
   };
 
   const handleBack = (): void => {
-    navigate('/welcome');
+    navigate('/welcome', { state: { step: 'GET_STARTED' } });
   };
 
   return (
