+++ src/client/components/welcome/GetStarted.tsx
