+++ src/client/components/welcome/Initial.tsx
